Move views to ./views
