+++ src/obsidian/augmentations/views/EditableFileView.d.ts
