+++ src/obsidian/augmentations/views/TextFileView.d.ts
