+++ src/obsidian/internals/views/AllPropertiesView.d.ts
