+++ src/obsidian/internals/views/BacklinkView.d.ts
@@ -1,4 +1,4 @@
-import type { InfoFileView } from "../types.js";
+import type { InfoFileView } from "../../types.js";
 /** @todo Documentation incomplete */
 export interface BacklinkView extends InfoFileView {
     /**
