+++ src/obsidian/internals/views/BrowserHistoryView.d.ts
