+++ src/obsidian/internals/views/BrowserView.d.ts
