+++ src/obsidian/internals/views/CanvasView.d.ts
