+++ src/obsidian/internals/views/EmptyView.d.ts
