+++ src/obsidian/internals/views/GraphView.d.ts
