+++ src/obsidian/internals/views/ImageView.d.ts
