+++ src/obsidian/internals/views/InfoFileView.d.ts
