+++ src/obsidian/internals/views/OutgoingLinkView.d.ts
@@ -1,4 +1,4 @@
-import type { InfoFileView } from "../types.js";
+import type { InfoFileView } from "../../types.js";
 /** @todo Documentation incomplete */
 export interface OutgoingLinkView extends InfoFileView {
     /**
