+++ src/obsidian/internals/views/OutlineView.d.ts
@@ -1,4 +1,4 @@
-import type { InfoFileView } from "../types.js";
+import type { InfoFileView } from "../../types.js";
 /** @todo Documentation incomplete */
 export interface OutlineView extends InfoFileView {
     findCorrespondingLeaf(): unknown;
