+++ src/obsidian/internals/views/PdfView.d.ts
