+++ src/obsidian/internals/views/ReleaseNotesView.d.ts
