+++ src/obsidian/internals/views/SearchView.d.ts
