+++ src/obsidian/internals/views/UnknownView.d.ts
@@ -1,4 +1,4 @@
-import type { EmptyView } from "../types.js";
+import type { EmptyView } from "../../types.js";
 /**
  * Views of plugins that have been deactivated become an UnknownView
  * @todo This is probably not the right term
