+++ src/obsidian/internals/views/VideoView.d.ts
